WIP parser: local array is not a global var
Why is this important?
Because the following code is not correct C:
...
int arr[5];
arr = {1,2,3,4};
...
You cannot assign to a previously declared array!

This commit is one step towards a fix for this error.
We are not assigning a global variable to a previously declared array
anymore but rather assigning the initializer value from the global variable to mentioned array.