parser: handle struct lowering
From https://lists.llvm.org/pipermail/cfe-dev/2013-January/027302.html:
When a function has a struct parameter or return type,
Clang may lower a struct parameter into...
	- a "byval" pointer (for a struct with several different members)
	- a vector (for a struct with a few float members)
	- two doubles (for a struct with two double members)
	- an i64 (for a struct with two i32 members)
... and possibly more variations.

But there is no information in the metadata about types created in this way.
Therefore, we detect the use of the struct type as an argument or return value
of a function and do not reconstruct these types from the metadata.