parser: fix an array size deduction from the DI.
e.g. `int array[g_var];`
generates:
```llvmir
!18 = !DILocalVariable(name: "__vla_expr0", scope: !12, type: !19, flags: DIFlagArtificial)
!19 = !DIBasicType(name: "unsigned long", size: 64, encoding: DW_ATE_unsigned)
!20 = !DILocation(line: 0, scope: !12)
!21 = !DILocalVariable(name: "array", scope: !12, file: !3, line: 2, type: !22)
!22 = !DICompositeType(tag: DW_TAG_array_type, baseType: !5, elements: !23)
!23 = !{!24}
```
We can savely set an array size to 0, because type_kinds will differ, and
the deduced type will not be used. I hope this is a valid argument.
By valid I mean safe.

In llvm2c arrayType does not have a constructor for non-intergral type.
So we cannot set the variable name as a size argument...