parser/negCmpInst: add a function to negate a CmpInst condition
I wrongly assume that llvm always generates positive loop conditions, e.g. if this is true, then iterate, but after -O3 optimisations, loop condition might be negated: if this is false, then continue to the next iteration.

This new function will reverse the loop condition if the situation described above occurs!
This might need a proper solution - this is hackery, because we do not
replace the whole expression, just a printed character e.g. "<" -> ">=".